requirements.txt
Pillow
streamlit
geopy